pe(certificate_table): make certificate data in `AttributeCertificate` a Cow<[u8]>
This is important/interesting in the context of writing certificates because
more than often we are not loading existing certificates from an owned [u8],
we are writing new certificates which comes from temporary references, we need
our AttributeCertificate entry to take ownership for them.

Cow can offer Borrowed | Owned in an transparent way.