pe(section_table): public constructor for `SectionTable`
This is a simple function to assemble a template section table with
zeros in the unpredictable fields, it's up to the caller to carry on the
fill procedure, e.g. a writer structure which can do the layout.

It can automatically set the virtual address and size based on another
binary to append it, only the on-disk offsets are left to the caller.