pe(utils): expose rva2offset to external consumers
Relative virtual address to disk offset is a classical operation
to perform for various things, it is likely this operation won't change
and if it changes, we can keep it around and deprecate it quite easily
in the next release.

An example for external consumers: compute debug directory data offsets from RVAs
to write a debug directory.