pe(data_directories): simplify writing data directories
Previously, we were handling the offset ourselves, but we can just use `gwrite_with`
to let the offset be moved.

We don't need to use the "read" offset, this is only a distraction and can get wrong results
sometimes.

We only need to write in the order and assume linearity.